# [IBD] batch block reads/writes during `AutoFile` serialization