@l0rinc 2025-01-24
> which claims a speedup in reindex-chainstate

> So the 9% improvement in unserializing seems to translate to roughly 1.2% total runtime improvement in -reindex-chainstate on my machine.

I can measure several runs to see if it makes any difference (to be able to claim anything reliable about a ~1% potential speedup), but the reads are only 27% faster, while the writes are 2.9x faster, so I'm not surprised that reindex doesn't shine compared to IBD. But I'm currently rerunning them with the rebased PR to make sure my measurements were realistic.
If you have any concrete measurement you'd like me to do, feel free to specify them.