@ryanofsky 2025-03-13
Code review d700b14491a27a6f4f9069dde86fe914784346ca

This approach seems reaonable but maybe it would be good to have a more general solution to providing buffered reads and writes instead of having these ad-hoc `std::vector<uint8_t>` / `SpanReader` / `DataStream` / `write_large` calls.

Ideally instead of:

```c++
std::vector<uint8_t> mem(blk_size);
filein >> Span{mem};
SpanReader(mem) >> TX_WITH_WITNESS(block);
// ...
fileout.write_large(DataStream(block_size) << TX_WITH_WITNESS(block));
```

Code could look more like:

```c++
BufferedReader{fileIn, blk_size} >> TX_WITH_WITNESS(block);
// ...
BufferedWriter{fileout, block_size} << TX_WITH_WITNESS(block);
```

But I guess as long as the buffers are only going to be used once and thrown away in this case (and not be fixed-size ring buffers) having buffer classes wouldn't substantively improve this code much aside from making it look nicer. Buffer stream classes would seem more useful in cases where more fields are serialized and sizes are not known in advance.

Just wanted to suggest this in case there are other places where serialization is slow and buffering could speed it up. We might want a more general solution if that is the case.