@l0rinc 2025-03-14
Thanks for the review @ryanofsky 

> would be good to have a more general solution to providing buffered reads and writes

That's what I had originally (with a fixed-size buffer, so was even more general than your suggestion), making the call-site-diff trivial: https://github.com/bitcoin/bitcoin/pull/31539.
It was suggested by @theuni that it may be simpler to serialize to buffer directly. It does seem simpler to reason about this one, especially since they basically had the same speed. What do you think?

What other places are candidates for similar buffering?