@l0rinc 2025-03-14
> since the BufferedReadOnlyFile class implemented there duplicated functionality of AutoFile, didn't support writing, and didn't work with any type of stream

There was a writer as well, but it resulted in some [weird CI failures](https://github.com/bitcoin/bitcoin/pull/31539#issuecomment-2612262087) and given this alternative I just closed the PR.

> disappointed if you initially had a general solution that made it easy to enable buffering any place and then were told that instead of using that

No, the suggested approach was better, that's why I kept this one. I'd be curious what other reviewers think, but I don't mind re-generalizing it (now or in a follow-up, when we find other usages).

> XORed file i/o is faster with buffering one place, probably we should expect XORed file i/o with buffering to be faster other places as well

It's not just the xor though, many small writes seemed to be slower in general, it seems faster if *we* buffer instead of the OS (likely because of locality, the file is a global resource that needs locking).

> don't think buffering is the type of thing that needs to be deployed on a case-by-case basis

I can get behind that, can you point me to other usages where we could use this?