@ryanofsky 2025-03-14
> I can get behind that, can you point me to other usages where we could use this?

I don't know, but if I were looking I'd just grep for places where AutoFile class is used. I don't have any real concerns here and I'm fine with the current approach. To the extent I do have concerns they are about readability and maintainability of code not performance, so I like an approach that makes it trivial to turn buffering on and off, which is I why thought BufferedReader / BufferedWriter stream adapters suggested in https://github.com/bitcoin/bitcoin/pull/31551#issuecomment-2722752125 might be a good idea. But when I posted that I didn't know a similar idea had already been tried before, so wouldn't suggest changing the approach again at this point without feedback from the other reviewers.