@l0rinc 2025-03-14
Thanks again for the hint @ryanofsky, I went through all changes again and [reworked the batching](https://github.com/bitcoin/bitcoin/compare/177a3bff143ae41a536b4d2b5a44a255b768bf32..5a1c2bd341c9586b090b6b40c20edb011cb616eb) - it's so much better this way and it includes both your idea and Cory's.

Additionally I fixed a few things after rebase:
* There was a  leftover `SaveBlock` from the previous refactor;
* Commits and docs still contained `build/src/benc/bench_bitcoin` - changed them to `build/bin/bench_bitcoin`;
* `BLOCK_SERIALIZATION_HEADER_SIZE` is used inline in the examples now and is too long - simplified it in a separate commit;
* Added `BufferedFileW` and `BufferedFileR`, encapsulating the buffering (and the new `AutoFile::write_large`) so that we can have a similar usage to what you suggested;
* I have also extracted the low-risk refactorings (which are needed for the batching later) to a separate commit to minimize the diffs in the risky commits.

The performance is same as before, the commits should be a lot simpler to review.