@l0rinc 2025-03-24
@ryanofsky, thanks again for the review.
Your suggestion seems to point back to the initial implementation of this PR, i.e. https://github.com/bitcoin/bitcoin/pull/31539/commits, where we didn't know the full size of the block, and read and wrote in chunks.
This current alternative was explicitly suggested by @theuni - and while it's a bit messy to always read the size, the resulting code is faster and seem to have fewer moving parts (e.g. doesn't reassign via the parameter, doesn't copy the write buffer, doesn't loop inside the buffers), it's why I went this way.
Maybe there's a middle ground by allocating max block size, ignoring the actual size - which would potentially simplify both approaches - would that be better than the current solutions? I'll investigate tomorrow...